Changed data types
All doubles are Doubles
timestamp now works.

+++ backend/src/main/java/dk/sdu/cps/stockwatch/builders/Builder.java
@@ -5,11 +5,11 @@ import dk.sdu.cps.stockwatch.model.Stock;
 import java.sql.Timestamp;
 
 public interface Builder {
-    StockBuilder setOpen(double open);
-    StockBuilder setHigh(double high);
-    StockBuilder setLow(double low);
-    StockBuilder setClose(double close);
-    StockBuilder setVolume(long volume);
+    StockBuilder setOpen(Double open);
+    StockBuilder setHigh(Double high);
+    StockBuilder setLow(Double low);
+    StockBuilder setClose(Double close);
+    StockBuilder setVolume(Double volume);
     StockBuilder setStock(Stock stock);
     StockBuilder setTimeStamp(Timestamp timestamp);
 }
